requirement.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.6
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
django-zipfile==0.4.0
djangorestframework==3.13.1
joblib==1.1.0
kombu==5.2.4
model-mommy==2.0.0
nltk==3.7
numpy==1.23.2
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyparsing==3.0.9
pyspellchecker==0.6.3
pytz==2022.1
redis==4.3.4
regex==2022.7.25
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
